refactor(jq): rename module
